Added some tests for ChicagoObjectDecoder
Not all of these tests pass.  The tests show my understanding of how
the decoder should be working, but my understanding might be off.

This is meant to start a conversation on what I'm not understanding
or how the code can be improved.

My thoughts are that the encoder and decoder should be mirrors
of each other.  Aggregating a STREAM_RESPONSE probably shouldn't
be in the decoder, that belongs in an Aggregator (Message->Message)

To handle the case where a STREAM_RESPONSE is spread over multiple
netty ByteBuf's there needs to be a token message sent to mark the
end of a stream.  I see the decoder trying to add such a marking
in the aggregated ChicagoMessage, but I don't see the encoder doing
anything like that, which makes proper decoding difficult.